gh-501 Make ClusterLockServiceTransactionHelper clearable
